// src/app/models/DirectorioModel.ts
export class DirectorioModel{
    SERINFODAT003_IdUsuario: Number;
    SERINFODAT003_Clave: Number;
    SERINFODAT003_NombreUsuario: string;
    SERINFODAT003_Extension: string;
    SERINFODAT003_Movil: string;
    SERINFODAT003_Casa: string;
    SERINFODAT003_Oficina: string;
    SERINFODAT003_Personalizado: string;
    SERINFODAT003_Correo : string;
    SERINFODAT003_Departamento : string;
    SERINFODAT003_IdLugar : number;
    SERINFODAT003_Admin : boolean; 
    SERINFODAT003_UsuarioInsert : string;
    SERINFODAT003_UsuarioBaja : string;
    SERINFODAT003_UsuarioModifico : string;
    SERINFODAT003_Estatus : boolean;
    
}